+++ sentry-ruby/lib/sentry/breadcrumb.rb
@@ -8,7 +8,7 @@ module Sentry
     # @return [String, nil]
     attr_accessor :category
     # @return [Hash, nil]
-    attr_accessor :data
+    attr_reader :data
     # @return [String, nil]
     attr_reader :level
     # @return [Time, Integer, nil]
@@ -26,7 +26,7 @@ module Sentry
     # @param type [String, nil]
     def initialize(category: nil, data: nil, message: nil, timestamp: nil, level: nil, type: nil)
       @category = category
-      @data = data || {}
+      self.data = data
       @timestamp = timestamp || Sentry.utc_now.to_i
       @type = type
       self.message = message
@@ -51,6 +51,14 @@ module Sentry
       @message = message && Utils::EncodingHelper.valid_utf_8?(message) ? message.byteslice(0..Event::MAX_MESSAGE_SIZE_IN_BYTES) : ""
     end
 
+    # Sanitizes the breadcrumb's arbitrary, user-supplied data so it doesn't
+    # carry a String with an invalid/non-UTF-8 encoding into JSON generation.
+    # @param data [Hash, nil]
+    # @return [void]
+    def data=(data)
+      @data = Utils::EncodingHelper.deep_encode_utf_8(data || {})
+    end
+
     # @param level [String]
     # @return [void]
     def level=(level) # needed to meet the Sentry spec
@@ -62,11 +70,6 @@ module Sentry
     def serialized_data
       begin
         ::JSON.parse(::JSON.generate(@data, max_nesting: MAX_NESTING))
-      rescue EncodingError, ::JSON::GeneratorError
-        # As of json 3.0, `JSON.generate` raises instead of warning when it
-        # encounters a String with an invalid/non-UTF-8 encoding (e.g. a
-        # BINARY-tagged String). Sanitize and retry once before giving up.
-        ::JSON.parse(::JSON.generate(Utils::EncodingHelper.deep_encode_utf_8(@data), max_nesting: MAX_NESTING))
       rescue Exception => e
         Sentry.sdk_logger.debug(LOGGER_PROGNAME) do
           <<~MSG